Show a helpful tips dialog
