+++ microscope/microscope.py
@@ -30,6 +30,7 @@ import importlib.resources
 import plotwindow
 import subscriber
 from channel_group import ChannelGroup
+from dialogs import AboutDialog, HelpDialog
 
 try:
     from _version import version as __version__
@@ -42,20 +43,6 @@ except ImportError:
 QCoreApplication.setOrganizationName("NIST Quantum Sensors")
 QCoreApplication.setOrganizationDomain("nist.gov")
 QCoreApplication.setApplicationName("Microscope")
-
-
-class AboutDialog(QtWidgets.QDialog):
-    def __init__(self, parent=None) -> None:
-        super().__init__(parent)
-
-        self.setWindowTitle("About Microscope")
-
-        layout = QtWidgets.QVBoxLayout()
-        layout.addWidget(QtWidgets.QLabel("<b>Microscope</b>"))
-        layout.addWidget(QtWidgets.QLabel(f"{parent.title} Python version"))
-        layout.addWidget(QtWidgets.QLabel("by Joe Fowler and NIST Boulder Labs, 2025"))
-        layout.addWidget(QtWidgets.QLabel("Not subject to copyright in the United States"))
-        self.setLayout(layout)
 
 
 def find_resource(filename):
@@ -80,7 +67,8 @@ class MainWindow(QtWidgets.QMainWindow):  # noqa: PLR0904
         PyQt5.uic.loadUi(find_resource("ui/microscope.ui"), self)
         self.setWindowTitle(title)
         self.title = title
-        self.actionAbout.triggered.connect(self.show_about)
+        self.actionAbout.triggered.connect(self.showAbout)
+        self.actionPlot_tips.triggered.connect(self.showPlotTips)
 
         self.statusLabel1 = QtWidgets.QLabel("")
         ss = "QLabel {{ color : black; size : 9}}"
@@ -101,8 +89,13 @@ class MainWindow(QtWidgets.QMainWindow):  # noqa: PLR0904
         self.addPlotsButton.clicked.connect(self.addPlotWindow)
 
     @pyqtSlot()
-    def show_about(self) -> None:
+    def showAbout(self) -> None:
         dialog = AboutDialog(self)
+        dialog.show()
+
+    @pyqtSlot()
+    def showPlotTips(self) -> None:
+        dialog = HelpDialog(self)
         dialog.show()
 
     @pyqtSlot()
